+++ src/main/java/org/aarboard/nextcloud/api/webdav/Folders.java
@@ -180,6 +180,17 @@ public class Folders extends AWebdavHandler{
     }
 
     /**
+     * method to rename/move folder
+     *
+     * @param oldPath path of the folder which should be renamed/moved
+     * @param newPath path of the folder which should be renamed/moved
+     * @param overwriteExisting overwrite if target already exists
+     */
+    public void renameFolder(String oldPath, String newPath, boolean overwriteExisting) {
+        renamePath(oldPath, newPath, overwriteExisting);
+    }
+    
+    /**
      * Downloads the folder at the specified remotePath to the rootDownloadDirPath
      *
      * @param remotePath the path in the nextcloud server with respect to the specific folder